requirements.txt
Django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
autopep8==2.3.1